Formatting fixes and cleanup
Also changed Vec3 -> vec3 and Quat -> quat (except for references to the Vec3 and Quat globals) for consistency. Some of the function references were still copy/pasted from the C++ code, so I removed the C++ bits (const keyword, & operator, glm:: namespace) as they weren't being shown in the rest of the document.